+++ next/lib/account-by-username-query.js
@@ -1,0 +1,13 @@
+import gql from 'graphql-tag';
+
+export default gql`
+  query publicAccount($username: String!) {
+    accountByUsername(username: $username) {
+      id
+      username
+      accountType
+      registered
+      profileIconId
+    }
+  }
+`;
